requirements.txt
flask
flask-pymongo
python-dotenv
google-generativeai
Werkzeug